+++ spec/integration/actions/create_chapter_spec.rb
@@ -22,7 +22,6 @@ RSpec.describe Spec::Support::Actions::CreateChapter do
   end
   let(:resource) do
     Cuprum::Rails::Resource.new(
-      collection:           repository.find_or_create(record_class: Chapter),
       permitted_attributes: %i[title],
       resource_class:       Chapter
     )